roedu-backend/requirements.txt
FastAPI==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
alembic==1.13.0
aiofiles==23.2.1
openai==1.3.7